requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
bidict==0.23.1
certifi==2025.4.26
click==8.1.8
fastapi==0.115.12
granian==2.2.5
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
numpy==2.2.6
packaging==24.2
platformdirs==4.3.7
psutil==7.0.0
psycopg==3.2.9
psycopg2-binary==2.9.10
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
python-decouple==3.8
python-dotenv==1.1.0
python-engineio==4.12.0
python-multipart==0.0.20
python-socketio==5.13.0
PyYAML==6.0.2
redis==5.2.1
reflex==0.7.13
reflex-hosting-cli==0.1.49
reflex-local-auth==0.3.0
rich==13.9.4
setuptools==80.8.0
shellingham==1.5.4
simple-websocket==1.1.0
sniffio==1.3.1
spglib==2.6.0
SQLAlchemy==2.0.40
sqlmodel==0.0.24
starlette==0.46.2
tabulate==0.9.0
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.2
watchfiles==1.0.5
wrapt==1.17.2
wsproto==1.2.0
